+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### Fix
+
+- When topic has messages without key or value, null is shown instead of an error
 ## 0.1.1
 
 ### Changed